Refactor SPI, add versioning
