+++ esp-hal/src/soc/esp32/mod.rs
@@ -22,6 +22,12 @@ pub(crate) fn i2s_sclk_frequency() -> u32 {
     clocks::pll_f160m_clk_frequency()
 }
 
+#[cfg(spi_master_driver_supported)]
+#[cfg_attr(not(feature = "unstable"), allow(unused))]
+pub(crate) fn spi_master_clock_source_frequency() -> u32 {
+    clocks::apb_clk_frequency()
+}
+
 pub(crate) unsafe fn configure_cpu_caches() {}
 
 pub(crate) fn pre_init() {}
