+++ esp-hal/src/soc/esp32c3/mod.rs
@@ -20,4 +20,10 @@ pub(crate) fn i2s_sclk_frequency() -> u32 {
     clocks::pll_160m_frequency()
 }
 
+#[cfg(spi_master_driver_supported)]
+#[cfg_attr(not(feature = "unstable"), allow(unused))]
+pub(crate) fn spi_master_clock_source_frequency() -> u32 {
+    clocks::apb_clk_frequency()
+}
+
 pub(crate) fn pre_init() {}
