+++ esp-hal/src/soc/esp32s3/mod.rs
@@ -23,6 +23,12 @@ pub(crate) fn i2s_sclk_frequency() -> u32 {
     clocks::pll_160m_frequency()
 }
 
+#[cfg(spi_master_driver_supported)]
+#[cfg_attr(not(feature = "unstable"), allow(unused))]
+pub(crate) fn spi_master_clock_source_frequency() -> u32 {
+    clocks::apb_clk_frequency()
+}
+
 #[unsafe(link_section = ".rwtext")]
 pub(crate) unsafe fn configure_cpu_caches() {
     // this is just the bare minimum we need to run code from flash
