+++ esp-hal/src/spi/master/mod.rs
@@ -34,52 +34,37 @@
 //! [`embedded-hal-bus`]: https://docs.rs/embedded-hal-bus/latest/embedded_hal_bus/spi/index.html
 //! [`embassy-embedded-hal`]: embassy_embedded_hal::shared_bus
 
-#[cfg(esp32)]
-use core::cell::Cell;
-use core::{
-    cell::UnsafeCell,
-    future::Future,
-    marker::PhantomData,
-    mem::MaybeUninit,
-    pin::Pin,
-    sync::atomic::{AtomicUsize, Ordering},
-    task::{Context, Poll},
-};
+use core::{marker::PhantomData, sync::atomic::Ordering};
 
 #[cfg(spi_master_supports_dma)]
 mod dma;
+mod low_level;
 
 #[instability::unstable]
 #[cfg(spi_master_supports_dma)]
 pub use dma::*;
 use embedded_hal::spi::SpiBus;
 use embedded_hal_async::spi::SpiBus as SpiBusAsync;
-use enumset::{EnumSet, EnumSetType, enum_set};
+use enumset::{EnumSet, EnumSetType};
+use low_level::{Driver, SpiWrapper};
+pub use low_level::{Info, Instance, QspiInstance, State};
 use procmacros::doc_replace;
 
 use super::{BitOrder, Error, Mode};
-#[cfg(spi_master_supports_dma)]
-use crate::RegisterToggle;
 use crate::{
     Async,
     Blocking,
     DriverMode,
-    asynch::AtomicWaker,
     gpio::{
         InputConfig,
-        InputSignal,
         NoPin,
         OutputConfig,
         OutputSignal,
         PinGuard,
         interconnect::{self, PeripheralInput, PeripheralOutput},
     },
-    handler,
     interrupt::InterruptHandler,
-    pac::spi2::RegisterBlock,
-    private::{self, DropGuard, Sealed},
-    ram,
-    system::PeripheralGuard,
+    private::Sealed,
     time::Rate,
 };
 
@@ -109,7 +94,7 @@ pub enum SpiInterrupt {
 }
 
 /// The size of the FIFO buffer for SPI
-const FIFO_SIZE: usize = if cfg!(esp32s2) { 72 } else { 64 };
+const FIFO_SIZE: usize = property!("spi_master.fifo_size") as usize;
 
 /// Padding byte for empty write transfers
 const EMPTY_WRITE_PAD: u8 = 0x00;
@@ -528,22 +513,7 @@ impl Config {
 
     fn clock_source_freq_hz(&self) -> Rate {
         // FIXME: take clock source into account
-        cfg_if::cfg_if! {
-            if #[cfg(esp32h2)] {
-                // ESP32-H2 is using PLL_48M_CLK source instead of APB_CLK
-                Rate::from_mhz(48)
-            } else if #[cfg(any(esp32c5, esp32c61))] {
-                // We select the 160MHz PLL as the clock source in the driver. There is a by-2 divider
-                // configured between the PLL and the SPI clock (spi2_clkm_div_num).
-                Rate::from_mhz(80) // clk_spi2_mst must be <= 80MHz
-            } else if #[cfg(esp32c6)] {
-                // We select the 80MHz PLL as the clock source in the driver
-                // FIXME we state that the default clock source is APB, which just isn't true
-                Rate::from_mhz(80)
-            } else {
-                Rate::from_hz(crate::soc::clocks::apb_clk_frequency())
-            }
-        }
+        Rate::from_hz(crate::soc::spi_master_clock_source_frequency())
     }
 
     fn recalculate(&self) -> Result<u32, ConfigError> {
@@ -679,65 +649,6 @@ impl core::fmt::Display for ConfigError {
             ConfigError::FrequencyOutOfRange => {
                 write!(f, "The requested frequency is not in the supported range")
             }
-        }
-    }
-}
-
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-struct SpiWrapper<'d> {
-    spi: AnySpi<'d>,
-    _guard: PeripheralGuard,
-}
-
-impl<'d> SpiWrapper<'d> {
-    fn new(spi: impl Instance + 'd) -> Self {
-        let p = spi.info().peripheral;
-        let this = Self {
-            spi: spi.degrade(),
-            _guard: PeripheralGuard::new(p),
-        };
-
-        // Initialize state
-        unsafe {
-            this.state()
-                .pins
-                .get()
-                .write(MaybeUninit::new(SpiPinGuard::new_unconnected()))
-        }
-
-        this
-    }
-
-    fn info(&self) -> &'static Info {
-        self.spi.info()
-    }
-
-    fn state(&self) -> &'static State {
-        self.spi.state()
-    }
-
-    fn disable_peri_interrupt_on_all_cores(&self) {
-        self.spi.disable_peri_interrupt_on_all_cores();
-    }
-
-    fn set_interrupt_handler(&self, handler: InterruptHandler) {
-        self.spi.set_interrupt_handler(handler);
-    }
-
-    fn pins(&mut self) -> &mut SpiPinGuard {
-        unsafe {
-            // SAFETY: we "own" the state, we are allowed to borrow it mutably
-            self.state().pins()
-        }
-    }
-}
-
-impl Drop for SpiWrapper<'_> {
-    fn drop(&mut self) {
-        unsafe {
-            // SAFETY: we "own" the state, we are allowed to deinit it
-            self.spi.state().deinit();
         }
     }
 }
@@ -1378,7 +1289,7 @@ where
         self.flush()?;
 
         cfg_if::cfg_if! {
-            if #[cfg(all(esp32, spi_address_workaround))] {
+            if #[cfg(all(spi_master_version = "1", spi_address_workaround))] {
                 let mut buffer = buffer;
                 let mut data_mode = data_mode;
                 let mut address = address;
@@ -1546,1069 +1457,6 @@ impl SpiBusAsync for Spi<'_, Async> {
     }
 }
 
-/// SPI peripheral instance.
-pub trait Instance: private::Sealed + any::Degrade {
-    #[doc(hidden)]
-    /// Returns the peripheral data and state describing this instance.
-    fn parts(&self) -> (&'static Info, &'static State);
-
-    /// Returns the peripheral data describing this instance.
-    #[doc(hidden)]
-    #[inline(always)]
-    fn info(&self) -> &'static Info {
-        self.parts().0
-    }
-
-    /// Returns the peripheral state for this instance.
-    #[doc(hidden)]
-    #[inline(always)]
-    fn state(&self) -> &'static State {
-        self.parts().1
-    }
-}
-
-/// Marker trait for QSPI-capable SPI peripherals.
-#[doc(hidden)]
-pub trait QspiInstance: Instance {}
-
-/// Peripheral data describing a particular SPI instance.
-#[doc(hidden)]
-#[non_exhaustive]
-#[allow(private_interfaces, reason = "Unstable details")]
-pub struct Info {
-    /// Pointer to the register block for this SPI instance.
-    ///
-    /// Use [Self::register_block] to access the register block.
-    pub register_block: *const RegisterBlock,
-
-    /// The system peripheral marker.
-    pub peripheral: crate::system::Peripheral,
-
-    /// Interrupt handler for the asynchronous operations.
-    pub async_handler: InterruptHandler,
-
-    /// SCLK signal.
-    pub sclk: OutputSignal,
-
-    /// Chip select signals.
-    pub cs: &'static [OutputSignal],
-
-    pub sio_inputs: &'static [InputSignal],
-    pub sio_outputs: &'static [OutputSignal],
-}
-
-impl Info {
-    fn cs(&self, n: usize) -> OutputSignal {
-        *unwrap!(self.cs.get(n), "CS{} is not defined", n)
-    }
-
-    fn opt_sio_input(&self, n: usize) -> Option<InputSignal> {
-        self.sio_inputs.get(n).copied()
-    }
-
-    fn opt_sio_output(&self, n: usize) -> Option<OutputSignal> {
-        self.sio_outputs.get(n).copied()
-    }
-
-    fn sio_input(&self, n: usize) -> InputSignal {
-        unwrap!(self.opt_sio_input(n), "SIO{} is not defined", n)
-    }
-
-    fn sio_output(&self, n: usize) -> OutputSignal {
-        unwrap!(self.opt_sio_output(n), "SIO{} is not defined", n)
-    }
-}
-
-struct Driver {
-    info: &'static Info,
-    state: &'static State,
-}
-
-// private implementation bits
-// FIXME: split this up into peripheral versions
-impl Driver {
-    /// Returns the register block for this SPI instance.
-    pub fn regs(&self) -> &RegisterBlock {
-        unsafe { &*self.info.register_block }
-    }
-
-    fn abort_transfer(&self) {
-        // Note(danielb): This method came later than DmaDriver::abort_transfer. That
-        // function works for DMA so I have left it unchanged, but does not work
-        // for CPU-controlled transfers on the ESP32. Toggling slave mode works on
-        // ESP32, but not on later chips.
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                self.regs().slave().toggle(|w, en| w.mode().bit(en));
-            } else {
-                self.configure_datalen(1, 1);
-            }
-        }
-        self.update();
-    }
-
-    /// Initialize for full-duplex 1 bit mode
-    fn init(&self) {
-        self.regs().user().modify(|_, w| {
-            w.usr_miso_highpart().clear_bit();
-            w.usr_mosi_highpart().clear_bit();
-            w.doutdin().set_bit();
-            w.usr_miso().set_bit();
-            w.usr_mosi().set_bit();
-            w.cs_hold().set_bit();
-            w.usr_dummy_idle().set_bit();
-            w.usr_addr().clear_bit();
-            w.usr_command().clear_bit()
-        });
-
-        #[cfg(not(any(esp32, esp32s2)))]
-        self.regs().clk_gate().modify(|_, w| {
-            w.clk_en().set_bit();
-            w.mst_clk_active().set_bit();
-            w.mst_clk_sel().set_bit()
-        });
-
-        #[cfg(soc_has_pcr)]
-        unsafe {
-            // use default clock source
-            crate::peripherals::PCR::regs()
-                .spi2_clkm_conf()
-                .modify(|_, w| {
-                    #[cfg(spi_master_has_clk_pre_div)]
-                    w.spi2_clkm_div_num().bits(1);
-                    w.spi2_clkm_sel().bits(1);
-                    w.spi2_clkm_en().set_bit()
-                });
-        }
-
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                self.regs().ctrl().modify(|_, w| w.wp().clear_bit());
-            } else {
-                self.regs().ctrl().modify(|_, w| {
-                    w.q_pol().clear_bit();
-                    w.d_pol().clear_bit();
-                    #[cfg(esp32s2)]
-                    w.wp().clear_bit();
-                    #[cfg(not(esp32s2))]
-                    w.hold_pol().clear_bit();
-                    w
-                });
-                self.regs().misc().write(|w| unsafe { w.bits(0) });
-            }
-        }
-
-        self.regs().slave().write(|w| unsafe { w.bits(0) });
-    }
-
-    #[cfg(not(esp32))]
-    fn init_spi_data_mode(
-        &self,
-        cmd_mode: DataMode,
-        address_mode: DataMode,
-        data_mode: DataMode,
-    ) -> Result<(), Error> {
-        self.regs().ctrl().modify(|_, w| {
-            w.fcmd_dual().bit(cmd_mode == DataMode::Dual);
-            w.fcmd_quad().bit(cmd_mode == DataMode::Quad);
-            w.faddr_dual().bit(address_mode == DataMode::Dual);
-            w.faddr_quad().bit(address_mode == DataMode::Quad);
-            w.fread_dual().bit(data_mode == DataMode::Dual);
-            w.fread_quad().bit(data_mode == DataMode::Quad)
-        });
-        self.regs().user().modify(|_, w| {
-            w.fwrite_dual().bit(data_mode == DataMode::Dual);
-            w.fwrite_quad().bit(data_mode == DataMode::Quad)
-        });
-        Ok(())
-    }
-
-    #[cfg(esp32)]
-    fn init_spi_data_mode(
-        &self,
-        cmd_mode: DataMode,
-        address_mode: DataMode,
-        data_mode: DataMode,
-    ) -> Result<(), Error> {
-        match cmd_mode {
-            DataMode::Single => (),
-            DataMode::SingleTwoDataLines => (),
-            // FIXME: more detailed error - Only 1-bit commands are supported.
-            _ => {
-                error!("Commands must be single bit wide");
-                return Err(Error::Unsupported);
-            }
-        }
-
-        match address_mode {
-            DataMode::Single | DataMode::SingleTwoDataLines => {
-                self.regs().ctrl().modify(|_, w| {
-                    w.fastrd_mode()
-                        .bit(matches!(data_mode, DataMode::Dual | DataMode::Quad));
-                    w.fread_dio().clear_bit();
-                    w.fread_qio().clear_bit();
-                    w.fread_dual().bit(data_mode == DataMode::Dual);
-                    w.fread_quad().bit(data_mode == DataMode::Quad)
-                });
-
-                self.regs().user().modify(|_, w| {
-                    w.fwrite_dio().clear_bit();
-                    w.fwrite_qio().clear_bit();
-                    w.fwrite_dual().bit(data_mode == DataMode::Dual);
-                    w.fwrite_quad().bit(data_mode == DataMode::Quad)
-                });
-            }
-            address_mode if address_mode == data_mode => {
-                self.regs().ctrl().modify(|_, w| {
-                    w.fastrd_mode()
-                        .bit(matches!(data_mode, DataMode::Dual | DataMode::Quad));
-                    w.fread_dio().bit(address_mode == DataMode::Dual);
-                    w.fread_qio().bit(address_mode == DataMode::Quad);
-                    w.fread_dual().clear_bit();
-                    w.fread_quad().clear_bit()
-                });
-
-                self.regs().user().modify(|_, w| {
-                    w.fwrite_dio().bit(address_mode == DataMode::Dual);
-                    w.fwrite_qio().bit(address_mode == DataMode::Quad);
-                    w.fwrite_dual().clear_bit();
-                    w.fwrite_quad().clear_bit()
-                });
-            }
-            _ => {
-                // FIXME: more detailed error - Unsupported combination of data-modes
-                error!("Address mode must be single bit wide or equal to the data mode");
-                return Err(Error::Unsupported);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Enable or disable listening for the given interrupts.
-    #[cfg_attr(not(feature = "unstable"), allow(dead_code))]
-    fn enable_listen(&self, interrupts: EnumSet<SpiInterrupt>, enable: bool) {
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                self.regs().slave().modify(|_, w| {
-                    for interrupt in interrupts {
-                        match interrupt {
-                            SpiInterrupt::TransferDone => w.trans_inten().bit(enable),
-                        };
-                    }
-                    w
-                });
-            } else if #[cfg(esp32s2)] {
-                self.regs().slave().modify(|_, w| {
-                    for interrupt in interrupts {
-                        match interrupt {
-                            SpiInterrupt::TransferDone => w.int_trans_done_en().bit(enable),
-                            SpiInterrupt::DmaSegmentedTransferDone => w.int_dma_seg_trans_en().bit(enable),
-                        };
-                    }
-                    w
-                });
-            } else {
-                self.regs().dma_int_ena().modify(|_, w| {
-                    for interrupt in interrupts {
-                        match interrupt {
-                            SpiInterrupt::TransferDone => w.trans_done().bit(enable),
-                            #[cfg(spi_master_has_dma_segmented_transfer)]
-                            SpiInterrupt::DmaSegmentedTransferDone => w.dma_seg_trans_done().bit(enable),
-                            #[cfg(spi_master_has_app_interrupts)]
-                            SpiInterrupt::App2 => w.app2().bit(enable),
-                            #[cfg(spi_master_has_app_interrupts)]
-                            SpiInterrupt::App1 => w.app1().bit(enable),
-                        };
-                    }
-                    w
-                });
-            }
-        }
-    }
-
-    /// Gets asserted interrupts
-    #[cfg_attr(not(feature = "unstable"), allow(dead_code))]
-    fn interrupts(&self) -> EnumSet<SpiInterrupt> {
-        let mut res = EnumSet::new();
-
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                if self.regs().slave().read().trans_done().bit() {
-                    res.insert(SpiInterrupt::TransferDone);
-                }
-            } else if #[cfg(esp32s2)] {
-                if self.regs().slave().read().trans_done().bit() {
-                    res.insert(SpiInterrupt::TransferDone);
-                }
-                if self.regs().hold().read().dma_seg_trans_done().bit() {
-                    res.insert(SpiInterrupt::DmaSegmentedTransferDone);
-                }
-            } else {
-                let ints = self.regs().dma_int_raw().read();
-
-                if ints.trans_done().bit() {
-                    res.insert(SpiInterrupt::TransferDone);
-                }
-                #[cfg(spi_master_has_dma_segmented_transfer)]
-                if ints.dma_seg_trans_done().bit() {
-                    res.insert(SpiInterrupt::DmaSegmentedTransferDone);
-                }
-                #[cfg(spi_master_has_app_interrupts)]
-                if ints.app2().bit() {
-                    res.insert(SpiInterrupt::App2);
-                }
-                #[cfg(spi_master_has_app_interrupts)]
-                if ints.app1().bit() {
-                    res.insert(SpiInterrupt::App1);
-                }
-            }
-        }
-
-        res
-    }
-
-    /// Resets asserted interrupts
-    fn clear_interrupts(&self, interrupts: EnumSet<SpiInterrupt>) {
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                for interrupt in interrupts {
-                    match interrupt {
-                        SpiInterrupt::TransferDone => {
-                            self.regs().slave().modify(|_, w| w.trans_done().clear_bit());
-                        }
-                    }
-                }
-            } else if #[cfg(esp32s2)] {
-                for interrupt in interrupts {
-                    match interrupt {
-                        SpiInterrupt::TransferDone => {
-                            self.regs().slave().modify(|_, w| w.trans_done().clear_bit());
-                        }
-
-                        SpiInterrupt::DmaSegmentedTransferDone => {
-                            self.regs()
-                                .hold()
-                                .modify(|_, w| w.dma_seg_trans_done().clear_bit());
-                        }
-                    }
-                }
-            } else {
-                self.regs().dma_int_clr().write(|w| {
-                    for interrupt in interrupts {
-                        match interrupt {
-                            SpiInterrupt::TransferDone => w.trans_done().clear_bit_by_one(),
-                            #[cfg(spi_master_has_dma_segmented_transfer)]
-                            SpiInterrupt::DmaSegmentedTransferDone => w.dma_seg_trans_done().clear_bit_by_one(),
-                            #[cfg(spi_master_has_app_interrupts)]
-                            SpiInterrupt::App2 => w.app2().clear_bit_by_one(),
-                            #[cfg(spi_master_has_app_interrupts)]
-                            SpiInterrupt::App1 => w.app1().clear_bit_by_one(),
-                        };
-                    }
-                    w
-                });
-            }
-        }
-    }
-
-    fn apply_config(&self, config: &Config) -> Result<(), ConfigError> {
-        config.validate()?;
-        self.ch_bus_freq(config)?;
-        self.set_bit_order(config.read_bit_order, config.write_bit_order);
-        self.set_data_mode(config.mode);
-        self.state
-            .min_async_transfer_size
-            .store(config.min_async_transfer_size, Ordering::Relaxed);
-
-        #[cfg(esp32)]
-        self.calculate_half_duplex_values(config);
-
-        Ok(())
-    }
-
-    #[cfg(esp32)]
-    fn calculate_half_duplex_values(&self, config: &Config) {
-        let f_apb = 80_000_000;
-        let source_freq_hz = match config.clock_source {
-            ClockSource::Apb => f_apb,
-        };
-
-        let clock_reg = self.regs().clock().read();
-        let eff_clk = if clock_reg.clk_equ_sysclk().bit_is_set() {
-            f_apb
-        } else {
-            let pre = clock_reg.clkdiv_pre().bits() as i32 + 1;
-            let n = clock_reg.clkcnt_n().bits() as i32 + 1;
-            f_apb / (pre * n)
-        };
-
-        let apbclk_khz = source_freq_hz / 1000;
-        // how many apb clocks a period has
-        let spiclk_apb_n = source_freq_hz / eff_clk;
-
-        // How many apb clocks the delay is, the 1 is to compensate in case
-        // ``input_delay_ns`` is rounded off. Change from esp-idf: we use
-        // `input_delay_ns` to also represent the GPIO matrix delay.
-        let input_delay_ns = 25; // TODO: allow configuring input delay.
-        let delay_apb_n = (1 + input_delay_ns) * apbclk_khz / 1000 / 1000;
-
-        let dummy_required = delay_apb_n / spiclk_apb_n;
-        let timing_miso_delay = if dummy_required > 0 {
-            // due to the clock delay between master and slave, there's a range in which
-            // data is random give MISO a delay if needed to make sure we
-            // sample at the time MISO is stable
-            Some(((dummy_required + 1) * spiclk_apb_n - delay_apb_n - 1) as u8)
-        } else if delay_apb_n * 4 <= spiclk_apb_n {
-            // if the dummy is not required, maybe we should also delay half a SPI clock if
-            // the data comes too early
-            None
-        } else {
-            Some(0)
-        };
-
-        self.state.esp32_hack.extra_dummy.set(dummy_required as u8);
-        self.state
-            .esp32_hack
-            .timing_miso_delay
-            .set(timing_miso_delay);
-    }
-
-    fn set_data_mode(&self, data_mode: Mode) {
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                let pin_reg = self.regs().pin();
-            } else {
-                let pin_reg = self.regs().misc();
-            }
-        };
-
-        pin_reg.modify(|_, w| {
-            w.ck_idle_edge()
-                .bit(matches!(data_mode, Mode::_2 | Mode::_3))
-        });
-        self.regs().user().modify(|_, w| {
-            w.ck_out_edge()
-                .bit(matches!(data_mode, Mode::_1 | Mode::_2))
-        });
-    }
-
-    fn ch_bus_freq(&self, bus_clock_config: &Config) -> Result<(), ConfigError> {
-        fn enable_clocks(_reg_block: &RegisterBlock, _enable: bool) {
-            #[cfg(not(any(esp32, esp32s2)))]
-            _reg_block.clk_gate().modify(|_, w| {
-                w.clk_en().bit(_enable);
-                w.mst_clk_active().bit(_enable);
-                w.mst_clk_sel().bit(true) // TODO: support XTAL clock source
-            });
-        }
-
-        // Change clock frequency
-        let raw = bus_clock_config.raw_clock_reg_value()?;
-
-        enable_clocks(self.regs(), false);
-        self.regs().clock().write(|w| unsafe { w.bits(raw) });
-        enable_clocks(self.regs(), true);
-
-        Ok(())
-    }
-
-    #[cfg(not(any(esp32, esp32c3, esp32s2)))]
-    fn set_bit_order(&self, read_order: BitOrder, write_order: BitOrder) {
-        let read_value = match read_order {
-            BitOrder::MsbFirst => 0,
-            BitOrder::LsbFirst => 1,
-        };
-        let write_value = match write_order {
-            BitOrder::MsbFirst => 0,
-            BitOrder::LsbFirst => 1,
-        };
-        self.regs().ctrl().modify(|_, w| unsafe {
-            w.rd_bit_order().bits(read_value);
-            w.wr_bit_order().bits(write_value);
-            w
-        });
-    }
-
-    #[cfg(any(esp32, esp32c3, esp32s2))]
-    fn set_bit_order(&self, read_order: BitOrder, write_order: BitOrder) {
-        let read_value = match read_order {
-            BitOrder::MsbFirst => false,
-            BitOrder::LsbFirst => true,
-        };
-        let write_value = match write_order {
-            BitOrder::MsbFirst => false,
-            BitOrder::LsbFirst => true,
-        };
-        self.regs().ctrl().modify(|_, w| {
-            w.rd_bit_order().bit(read_value);
-            w.wr_bit_order().bit(write_value);
-            w
-        });
-    }
-
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn fill_fifo(&self, chunk: &[u8]) {
-        // TODO: replace with `array_chunks` and `from_le_bytes`
-        let mut c_iter = chunk.chunks_exact(4);
-        let mut w_iter = self.regs().w_iter();
-        for c in c_iter.by_ref() {
-            if let Some(w_reg) = w_iter.next() {
-                let word = (c[0] as u32)
-                    | ((c[1] as u32) << 8)
-                    | ((c[2] as u32) << 16)
-                    | ((c[3] as u32) << 24);
-                w_reg.write(|w| w.buf().set(word));
-            }
-        }
-        let rem = c_iter.remainder();
-        if !rem.is_empty()
-            && let Some(w_reg) = w_iter.next()
-        {
-            let word = match rem.len() {
-                3 => (rem[0] as u32) | ((rem[1] as u32) << 8) | ((rem[2] as u32) << 16),
-                2 => (rem[0] as u32) | ((rem[1] as u32) << 8),
-                1 => rem[0] as u32,
-                _ => unreachable!(),
-            };
-            w_reg.write(|w| w.buf().set(word));
-        }
-    }
-
-    /// Write bytes to SPI.
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn write_one(&self, words: &[u8]) -> Result<(), Error> {
-        if words.len() > FIFO_SIZE {
-            return Err(Error::FifoSizeExeeded);
-        }
-        self.configure_datalen(0, words.len());
-        self.fill_fifo(words);
-        self.start_operation();
-        Ok(())
-    }
-
-    /// Write bytes to SPI.
-    ///
-    /// This function will return before all bytes of the last chunk to transmit
-    /// have been sent to the wire. If you must ensure that the whole
-    /// messages was written correctly, use [`Self::flush`].
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn write(&self, words: &[u8]) -> Result<(), Error> {
-        for chunk in words.chunks(FIFO_SIZE) {
-            self.flush()?;
-            self.write_one(chunk)?;
-        }
-        Ok(())
-    }
-
-    /// Write bytes to SPI.
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    async fn write_async(&self, words: &[u8]) -> Result<(), Error> {
-        for chunk in words.chunks(FIFO_SIZE) {
-            self.write_one(chunk)?;
-            self.flush_async().await;
-        }
-        Ok(())
-    }
-
-    /// Read bytes from SPI.
-    ///
-    /// Sends out a stuffing byte for every byte to read. This function doesn't
-    /// perform flushing. If you want to read the response to something you
-    /// have written before, consider using [`Self::transfer`] instead.
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn read(&self, words: &mut [u8]) -> Result<(), Error> {
-        let empty_array = [EMPTY_WRITE_PAD; FIFO_SIZE];
-
-        for chunk in words.chunks_mut(FIFO_SIZE) {
-            self.write_one(&empty_array[0..chunk.len()])?;
-            self.flush()?;
-            self.read_from_fifo(chunk)?;
-        }
-        Ok(())
-    }
-
-    /// Read bytes from SPI.
-    ///
-    /// Sends out a stuffing byte for every byte to read. If you want to read
-    /// the response to something you have written before, consider using
-    /// [`Self::transfer`] instead.
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    async fn read_async(&self, words: &mut [u8]) -> Result<(), Error> {
-        let empty_array = [EMPTY_WRITE_PAD; FIFO_SIZE];
-
-        for chunk in words.chunks_mut(FIFO_SIZE) {
-            self.write_one(&empty_array[0..chunk.len()])?;
-            self.flush_async().await;
-            self.read_from_fifo(chunk)?;
-        }
-        Ok(())
-    }
-
-    /// Read received bytes from SPI FIFO.
-    ///
-    /// Copies the contents of the SPI receive FIFO into `words`. This function
-    /// doesn't perform any data transfer. If you want to read the response to
-    /// something you have written before, consider using [`Self::transfer`]
-    /// instead.
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn read_from_fifo(&self, words: &mut [u8]) -> Result<(), Error> {
-        if words.len() > FIFO_SIZE {
-            return Err(Error::FifoSizeExeeded);
-        }
-
-        for (chunk, w_reg) in words.chunks_mut(4).zip(self.regs().w_iter()) {
-            let reg_val = w_reg.read().bits();
-            let bytes = reg_val.to_le_bytes();
-
-            let len = chunk.len();
-            chunk.copy_from_slice(&bytes[..len]);
-        }
-
-        Ok(())
-    }
-
-    fn busy(&self) -> bool {
-        self.regs().cmd().read().usr().bit_is_set()
-    }
-
-    // Check if the bus is busy and if it is wait for it to be idle
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn flush_async(&self) -> impl Future<Output = ()> {
-        SpiFuture { driver: self }
-    }
-
-    // Check if the bus is busy and if it is wait for it to be idle
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn flush(&self) -> Result<(), Error> {
-        while self.busy() {
-            // wait for bus to be clear
-        }
-        Ok(())
-    }
-
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn transfer_in_place(&self, words: &mut [u8]) -> Result<(), Error> {
-        for chunk in words.chunks_mut(FIFO_SIZE) {
-            self.write_one(chunk)?;
-            self.flush()?;
-            self.read_from_fifo(chunk)?;
-        }
-
-        Ok(())
-    }
-
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn transfer(&self, read: &mut [u8], write: &[u8]) -> Result<(), Error> {
-        let mut write_from = 0;
-        let mut read_from = 0;
-
-        loop {
-            // How many bytes we write in this chunk
-            let write_inc = core::cmp::min(FIFO_SIZE, write.len() - write_from);
-            // How many bytes we read in this chunk
-            let read_inc = core::cmp::min(FIFO_SIZE, read.len() - read_from);
-
-            if (write_inc == 0) && (read_inc == 0) {
-                break;
-            }
-
-            self.flush()?;
-
-            if write_inc < read_inc {
-                // Read more than we write, must pad writing part with zeros
-                let mut empty = [EMPTY_WRITE_PAD; FIFO_SIZE];
-                empty[0..write_inc].copy_from_slice(&write[write_from..][..write_inc]);
-                self.write_one(&empty[..read_inc])?;
-            } else {
-                self.write_one(&write[write_from..][..write_inc])?;
-            }
-
-            if read_inc > 0 {
-                self.flush()?;
-                self.read_from_fifo(&mut read[read_from..][..read_inc])?;
-            }
-
-            write_from += write_inc;
-            read_from += read_inc;
-        }
-        Ok(())
-    }
-
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    async fn transfer_in_place_async(&self, words: &mut [u8]) -> Result<(), Error> {
-        for chunk in words.chunks_mut(FIFO_SIZE) {
-            // Cut the transfer short if the future is dropped. We'll block for a short
-            // while to ensure the peripheral is idle.
-            let cancel_on_drop = DropGuard::new((), |_| {
-                self.abort_transfer();
-                while self.busy() {}
-            });
-            let res = self.write_one(chunk);
-            self.flush_async().await;
-            cancel_on_drop.defuse();
-            res?;
-
-            self.read_from_fifo(chunk)?;
-        }
-
-        Ok(())
-    }
-
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    async fn transfer_async(&self, read: &mut [u8], write: &[u8]) -> Result<(), Error> {
-        let mut write_from = 0;
-        let mut read_from = 0;
-
-        loop {
-            // How many bytes we write in this chunk
-            let write_inc = core::cmp::min(FIFO_SIZE, write.len() - write_from);
-            // How many bytes we read in this chunk
-            let read_inc = core::cmp::min(FIFO_SIZE, read.len() - read_from);
-
-            if (write_inc == 0) && (read_inc == 0) {
-                break;
-            }
-
-            self.flush_async().await;
-
-            if write_inc < read_inc {
-                // Read more than we write, must pad writing part with zeros
-                let mut empty = [EMPTY_WRITE_PAD; FIFO_SIZE];
-                empty[0..write_inc].copy_from_slice(&write[write_from..][..write_inc]);
-                self.write_one(&empty[..read_inc])?;
-            } else {
-                self.write_one(&write[write_from..][..write_inc])?;
-            }
-
-            // Preserve previous semantics - see https://github.com/esp-rs/esp-hal/issues/5257
-            self.flush_async().await;
-            if read_inc > 0 {
-                self.read_from_fifo(&mut read[read_from..][..read_inc])?;
-            }
-
-            write_from += write_inc;
-            read_from += read_inc;
-        }
-        Ok(())
-    }
-
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn start_operation(&self) {
-        self.update();
-        self.clear_interrupts(SpiInterrupt::TransferDone.into());
-        self.regs().cmd().modify(|_, w| w.usr().set_bit());
-    }
-
-    fn setup_full_duplex(&self) -> Result<(), Error> {
-        self.regs().user().modify(|_, w| {
-            w.usr_miso().set_bit();
-            w.usr_mosi().set_bit();
-            w.doutdin().set_bit();
-            w.usr_dummy().clear_bit();
-            w.sio().clear_bit()
-        });
-
-        self.init_spi_data_mode(
-            DataMode::SingleTwoDataLines,
-            DataMode::SingleTwoDataLines,
-            DataMode::SingleTwoDataLines,
-        )?;
-
-        // For full-duplex, we don't need compensation according to esp-idf.
-        #[cfg(esp32)]
-        self.regs().ctrl2().modify(|_, w| unsafe {
-            w.miso_delay_mode().bits(0);
-            w.miso_delay_num().bits(0)
-        });
-
-        Ok(())
-    }
-
-    #[expect(clippy::too_many_arguments)]
-    fn setup_half_duplex(
-        &self,
-        is_write: bool,
-        cmd: Command,
-        address: Address,
-        dummy_idle: bool,
-        dummy: u8,
-        no_mosi_miso: bool,
-        data_mode: DataMode,
-    ) -> Result<(), Error> {
-        self.init_spi_data_mode(cmd.mode(), address.mode(), data_mode)?;
-
-        #[cfg(esp32)]
-        let mut dummy = dummy;
-
-        #[cfg(esp32)]
-        self.regs().ctrl2().modify(|_, w| {
-            let mut delay_mode = 0;
-            let mut delay_num = 0;
-
-            if !is_write {
-                // Values are set up in apply_config
-                let timing_miso_delay = self.state.esp32_hack.timing_miso_delay.get();
-                let extra_dummy = self.state.esp32_hack.extra_dummy.get();
-                dummy += extra_dummy;
-
-                if let Some(delay) = timing_miso_delay {
-                    delay_num = if extra_dummy > 0 { delay } else { 0 };
-                } else {
-                    let out_edge = self.regs().user().read().ck_out_edge().bit_is_set();
-                    // SPI modes 1 and 2 need delay mode 1 according to esp-idf.
-                    delay_mode = if out_edge { 1 } else { 2 };
-                }
-            }
-
-            unsafe {
-                w.miso_delay_mode().bits(delay_mode);
-                w.miso_delay_num().bits(delay_num)
-            }
-        });
-
-        let reg_block = self.regs();
-        reg_block.user().modify(|_, w| {
-            w.usr_miso_highpart().clear_bit();
-            w.usr_mosi_highpart().clear_bit();
-            // This bit tells the hardware whether we use Single or SingleTwoDataLines
-            w.sio().bit(data_mode == DataMode::Single);
-            w.doutdin().clear_bit();
-            w.usr_miso().bit(!is_write && !no_mosi_miso);
-            w.usr_mosi().bit(is_write && !no_mosi_miso);
-            w.cs_hold().set_bit();
-            w.usr_dummy_idle().bit(dummy_idle);
-            w.usr_dummy().bit(dummy != 0);
-            w.usr_addr().bit(!address.is_none());
-            w.usr_command().bit(!cmd.is_none())
-        });
-
-        // FIXME why is clock config even here?
-        #[cfg(not(any(esp32, esp32s2)))]
-        reg_block.clk_gate().modify(|_, w| {
-            w.clk_en().set_bit();
-            w.mst_clk_active().set_bit();
-            w.mst_clk_sel().set_bit()
-        });
-
-        #[cfg(soc_has_pcr)]
-        // use default clock source
-        crate::peripherals::PCR::regs()
-            .spi2_clkm_conf()
-            .modify(|_, w| unsafe { w.spi2_clkm_sel().bits(1) });
-
-        #[cfg(not(esp32))]
-        reg_block.misc().write(|w| unsafe { w.bits(0) });
-
-        reg_block.slave().write(|w| unsafe { w.bits(0) });
-
-        self.update();
-
-        // set cmd, address, dummy cycles
-        self.set_up_common_phases(cmd, address, dummy);
-
-        Ok(())
-    }
-
-    fn set_up_common_phases(&self, cmd: Command, address: Address, dummy: u8) {
-        let reg_block = self.regs();
-        if !cmd.is_none() {
-            reg_block.user2().modify(|_, w| unsafe {
-                w.usr_command_bitlen().bits((cmd.width() - 1) as u8);
-                w.usr_command_value().bits(cmd.value())
-            });
-        }
-
-        if !address.is_none() {
-            reg_block
-                .user1()
-                .modify(|_, w| unsafe { w.usr_addr_bitlen().bits((address.width() - 1) as u8) });
-
-            let addr = address.value() << (32 - address.width());
-            #[cfg(not(esp32))]
-            reg_block
-                .addr()
-                .write(|w| unsafe { w.usr_addr_value().bits(addr) });
-            #[cfg(esp32)]
-            reg_block.addr().write(|w| unsafe { w.bits(addr) });
-        }
-
-        if dummy > 0 {
-            reg_block
-                .user1()
-                .modify(|_, w| unsafe { w.usr_dummy_cyclelen().bits(dummy - 1) });
-        }
-    }
-
-    fn update(&self) {
-        cfg_if::cfg_if! {
-            if #[cfg(esp32p4)] {
-                // P4 PAC: cmd().update() is write-only -- no `read()` accessor,
-                // so we can't poll for the bit clearing like the other chips.
-                // Spin for ~10 cycles to let the hardware latch instead.
-                // TODO: file an esp-pacs issue/PR so the P4 SVD marks the SPI
-                // cmd().update field readable like other chips. Once that lands
-                // this branch can collapse into the general `else` arm below.
-                let reg_block = self.regs();
-                reg_block.cmd().modify(|_, w| w.update().set_bit());
-                for _ in 0..10 { core::hint::spin_loop(); }
-            } else if #[cfg(not(any(esp32, esp32s2)))] {
-                let reg_block = self.regs();
-
-                reg_block.cmd().modify(|_, w| w.update().set_bit());
-
-                while reg_block.cmd().read().update().bit_is_set() {
-                    // wait
-                }
-            } else {
-                // Doesn't seem to be needed for ESP32 and ESP32-S2
-            }
-        }
-    }
-
-    fn configure_datalen(&self, rx_len_bytes: usize, tx_len_bytes: usize) {
-        let reg_block = self.regs();
-
-        let rx_len = rx_len_bytes as u32 * 8;
-        let tx_len = tx_len_bytes as u32 * 8;
-
-        let rx_len = rx_len.saturating_sub(1);
-        let tx_len = tx_len.saturating_sub(1);
-
-        cfg_if::cfg_if! {
-            if #[cfg(esp32)] {
-                let len = rx_len.max(tx_len);
-                reg_block
-                    .mosi_dlen()
-                    .write(|w| unsafe { w.usr_mosi_dbitlen().bits(len) });
-
-                reg_block
-                    .miso_dlen()
-                    .write(|w| unsafe { w.usr_miso_dbitlen().bits(len) });
-            } else if #[cfg(esp32s2)] {
-                reg_block
-                    .mosi_dlen()
-                    .write(|w| unsafe { w.usr_mosi_dbitlen().bits(tx_len) });
-
-                reg_block
-                    .miso_dlen()
-                    .write(|w| unsafe { w.usr_miso_dbitlen().bits(rx_len) });
-            } else {
-                reg_block
-                    .ms_dlen()
-                    .write(|w| unsafe { w.ms_data_bitlen().bits(rx_len.max(tx_len)) });
-            }
-
-        }
-    }
-}
-
-impl PartialEq for Info {
-    fn eq(&self, other: &Self) -> bool {
-        core::ptr::eq(self.register_block, other.register_block)
-    }
-}
-
-unsafe impl Sync for Info {}
-
-for_each_spi_master! {
-    ($peri:ident, $sys:ident, $sclk:ident [$($cs:ident),+] [$($sio:ident),*] $(, $is_qspi:tt)?) => {
-        impl Instance for crate::peripherals::$peri<'_> {
-            #[inline(always)]
-            fn parts(&self) -> (&'static Info, &'static State) {
-                #[handler]
-                #[ram]
-                fn irq_handler() {
-                    handle_async(&INFO, &STATE)
-                }
-
-                static INFO: Info = Info {
-                    register_block: crate::peripherals::$peri::regs(),
-                    peripheral: crate::system::Peripheral::$sys,
-                    async_handler: irq_handler,
-                    sclk: OutputSignal::$sclk,
-                    cs: &[$(OutputSignal::$cs),+],
-                    sio_inputs: &[$(InputSignal::$sio),*],
-                    sio_outputs: &[$(OutputSignal::$sio),*],
-                };
-
-                static STATE: State = State {
-                    waker: AtomicWaker::new(),
-                    pins: UnsafeCell::new(MaybeUninit::uninit()),
-                    min_async_transfer_size: AtomicUsize::new(0),
-
-                    #[cfg(esp32)]
-                    esp32_hack: Esp32Hack {
-                        timing_miso_delay: Cell::new(None),
-                        extra_dummy: Cell::new(0),
-                    },
-                };
-
-                (&INFO, &STATE)
-            }
-        }
-
-        $(
-            // If the extra pins are set, implement QspiInstance
-            $crate::ignore!($is_qspi);
-            impl QspiInstance for crate::peripherals::$peri<'_> {}
-        )?
-    };
-}
-
-impl QspiInstance for AnySpi<'_> {}
-
-#[doc(hidden)]
-pub struct State {
-    waker: AtomicWaker,
-    pins: UnsafeCell<MaybeUninit<SpiPinGuard>>,
-    min_async_transfer_size: AtomicUsize,
-
-    #[cfg(esp32)]
-    esp32_hack: Esp32Hack,
-}
-
-impl State {
-    // Syntactic helper to get a mutable reference to the pin guard.
-    //
-    // Intended to be called in `SpiWrapper::pins` only
-    //
-    // # Safety
-    //
-    // The caller must ensure that Rust's aliasing rules are upheld.
-    #[allow(
-        clippy::mut_from_ref,
-        reason = "Safety requirements ensure this is okay"
-    )]
-    unsafe fn pins(&self) -> &mut SpiPinGuard {
-        unsafe { (&mut *self.pins.get()).assume_init_mut() }
-    }
-
-    unsafe fn deinit(&self) {
-        unsafe {
-            let mut old = self.pins.get().replace(MaybeUninit::uninit());
-            old.assume_init_drop();
-        }
-    }
-}
-
-#[cfg(esp32)]
-struct Esp32Hack {
-    timing_miso_delay: Cell<Option<u8>>,
-    extra_dummy: Cell<u8>,
-}
-
-unsafe impl Sync for State {}
-
-#[ram]
-fn handle_async(info: &'static Info, state: &'static State) {
-    let driver = Driver { info, state };
-    if driver.interrupts().contains(SpiInterrupt::TransferDone) {
-        driver.enable_listen(SpiInterrupt::TransferDone.into(), false);
-        state.waker.wake();
-    }
-}
-
 /// SPI data mode
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
@@ -2637,6 +1485,8 @@ crate::any_peripheral! {
     }
 }
 
+impl QspiInstance for AnySpi<'_> {}
+
 impl Instance for AnySpi<'_> {
     #[inline]
     fn parts(&self) -> (&'static Info, &'static State) {
@@ -2656,36 +1506,5 @@ impl AnySpi<'_> {
     fn set_interrupt_handler(&self, handler: InterruptHandler) {
         self.disable_peri_interrupt_on_all_cores();
         self.bind_peri_interrupt(handler);
-    }
-}
-
-struct SpiFuture<'a> {
-    driver: &'a Driver,
-}
-
-impl SpiFuture<'_> {
-    const EVENTS: EnumSet<SpiInterrupt> = enum_set!(SpiInterrupt::TransferDone);
-}
-
-impl Future for SpiFuture<'_> {
-    type Output = ();
-
-    #[cfg_attr(place_spi_master_driver_in_ram, ram)]
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        if self.driver.busy() {
-            self.driver.state.waker.register(cx.waker());
-            self.driver.enable_listen(Self::EVENTS, true);
-
-            Poll::Pending
-        } else {
-            self.driver.clear_interrupts(Self::EVENTS);
-            Poll::Ready(())
-        }
-    }
-}
-
-impl Drop for SpiFuture<'_> {
-    fn drop(&mut self) {
-        self.driver.enable_listen(Self::EVENTS, false);
     }
 }
