+++ esp-hal/src/uart/compat.rs
@@ -1,16 +1,7 @@
-use super::{
-    Async,
-    Blocking,
-    Config,
-    ConfigError,
-    DriverMode,
-    IoError,
-    RxError,
-    TxError,
-    Uart,
-    UartRx,
-    UartTx,
-};
+#[cfg(feature = "unstable")]
+use super::{Async, Config, ConfigError, IoError, RxError, TxError, UartRx};
+use super::{Blocking, DriverMode, Uart, UartTx};
+#[cfg(feature = "unstable")]
 use crate::interrupt::InterruptHandler;
 
 #[instability::unstable]
