+++ esp-metadata-generated/src/_build_script_utils.rs
@@ -397,6 +397,9 @@ impl Chip {
                     "soc_has_clock_node_uart_baud_rate_generator",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"1\"",
+                    "spi_master_fifo_size=\"64\"",
+                    "spi_master_bit_order_is_bool",
                     "spi_master_supports_dma",
                     "spi_slave_supports_dma",
                     "timergroup_timg_has_timer1",
@@ -607,6 +610,9 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_uart_baud_rate_generator",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"1\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
+                    "cargo:rustc-cfg=spi_master_bit_order_is_bool",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_slave_supports_dma",
                     "cargo:rustc-cfg=timergroup_timg_has_timer1",
@@ -928,6 +934,8 @@ impl Chip {
                     "soc_has_clock_node_uart_baud_rate_generator",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
                     "spi_master_supports_dma",
                     "spi_master_has_app_interrupts",
                     "spi_master_has_dma_segmented_transfer",
@@ -1090,6 +1098,8 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_uart_baud_rate_generator",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_app_interrupts",
                     "cargo:rustc-cfg=spi_master_has_dma_segmented_transfer",
@@ -1411,6 +1421,9 @@ impl Chip {
                     "soc_has_clock_node_uart_baud_rate_generator",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
+                    "spi_master_bit_order_is_bool",
                     "spi_master_supports_dma",
                     "spi_master_has_app_interrupts",
                     "spi_master_has_dma_segmented_transfer",
@@ -1627,6 +1640,9 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_uart_baud_rate_generator",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
+                    "cargo:rustc-cfg=spi_master_bit_order_is_bool",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_app_interrupts",
                     "cargo:rustc-cfg=spi_master_has_dma_segmented_transfer",
@@ -1998,10 +2014,13 @@ impl Chip {
                     "soc_has_clock_node_parl_io_tx_clock",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
                     "spi_master_supports_dma",
                     "spi_master_has_app_interrupts",
                     "spi_master_has_dma_segmented_transfer",
                     "spi_master_has_clk_pre_div",
+                    "spi_master_dma_can_access_flash",
                     "timergroup_timg_has_divcnt_rst",
                     "timergroup_rc_fast_calibration_is_set",
                     "uart_ram_size=\"128\"",
@@ -2261,10 +2280,13 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_parl_io_tx_clock",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_app_interrupts",
                     "cargo:rustc-cfg=spi_master_has_dma_segmented_transfer",
                     "cargo:rustc-cfg=spi_master_has_clk_pre_div",
+                    "cargo:rustc-cfg=spi_master_dma_can_access_flash",
                     "cargo:rustc-cfg=timergroup_timg_has_divcnt_rst",
                     "cargo:rustc-cfg=timergroup_rc_fast_calibration_is_set",
                     "cargo:rustc-cfg=uart_ram_size=\"128\"",
@@ -2687,6 +2709,8 @@ impl Chip {
                     "soc_has_clock_node_parl_io_tx_clock",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
                     "spi_master_supports_dma",
                     "spi_master_has_app_interrupts",
                     "spi_master_has_dma_segmented_transfer",
@@ -2975,6 +2999,8 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_parl_io_tx_clock",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_app_interrupts",
                     "cargo:rustc-cfg=spi_master_has_dma_segmented_transfer",
@@ -3325,10 +3351,13 @@ impl Chip {
                     "soc_has_clock_node_timg_wdt_clock",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
                     "spi_master_supports_dma",
                     "spi_master_has_app_interrupts",
                     "spi_master_has_dma_segmented_transfer",
                     "spi_master_has_clk_pre_div",
+                    "spi_master_dma_can_access_flash",
                     "timergroup_timg_has_divcnt_rst",
                     "timergroup_rc_fast_calibration_is_set",
                     "uart_ram_size=\"128\"",
@@ -3528,10 +3557,13 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_timg_wdt_clock",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_app_interrupts",
                     "cargo:rustc-cfg=spi_master_has_dma_segmented_transfer",
                     "cargo:rustc-cfg=spi_master_has_clk_pre_div",
+                    "cargo:rustc-cfg=spi_master_dma_can_access_flash",
                     "cargo:rustc-cfg=timergroup_timg_has_divcnt_rst",
                     "cargo:rustc-cfg=timergroup_rc_fast_calibration_is_set",
                     "cargo:rustc-cfg=uart_ram_size=\"128\"",
@@ -3923,6 +3955,8 @@ impl Chip {
                     "soc_has_clock_node_parl_io_tx_clock",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
                     "spi_master_supports_dma",
                     "spi_master_has_app_interrupts",
                     "spi_master_has_dma_segmented_transfer",
@@ -4175,6 +4209,8 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_parl_io_tx_clock",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_app_interrupts",
                     "cargo:rustc-cfg=spi_master_has_dma_segmented_transfer",
@@ -4452,6 +4488,8 @@ impl Chip {
                     "soc_has_clock_node_timg_wdt_clock",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
                     "timergroup_timg_has_divcnt_rst",
                     "timergroup_rc_fast_calibration_is_set",
                     "uart_ram_size=\"128\"",
@@ -4615,6 +4653,8 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_timg_wdt_clock",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
                     "cargo:rustc-cfg=timergroup_timg_has_divcnt_rst",
                     "cargo:rustc-cfg=timergroup_rc_fast_calibration_is_set",
                     "cargo:rustc-cfg=uart_ram_size=\"128\"",
@@ -5075,6 +5115,9 @@ impl Chip {
                     "soc_has_clock_node_uart_mem_clock",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"2\"",
+                    "spi_master_fifo_size=\"72\"",
+                    "spi_master_bit_order_is_bool",
                     "spi_master_supports_dma",
                     "spi_master_has_octal",
                     "spi_master_has_dma_segmented_transfer",
@@ -5299,6 +5342,9 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_uart_mem_clock",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"2\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"72\"",
+                    "cargo:rustc-cfg=spi_master_bit_order_is_bool",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_octal",
                     "cargo:rustc-cfg=spi_master_has_dma_segmented_transfer",
@@ -5752,6 +5798,8 @@ impl Chip {
                     "soc_has_clock_node_uart_mem_clock",
                     "has_dram_region",
                     "has_dram2_uninit_region",
+                    "spi_master_version=\"3\"",
+                    "spi_master_fifo_size=\"64\"",
                     "spi_master_supports_dma",
                     "spi_master_has_octal",
                     "spi_master_has_app_interrupts",
@@ -6014,6 +6062,8 @@ impl Chip {
                     "cargo:rustc-cfg=soc_has_clock_node_uart_mem_clock",
                     "cargo:rustc-cfg=has_dram_region",
                     "cargo:rustc-cfg=has_dram2_uninit_region",
+                    "cargo:rustc-cfg=spi_master_version=\"3\"",
+                    "cargo:rustc-cfg=spi_master_fifo_size=\"64\"",
                     "cargo:rustc-cfg=spi_master_supports_dma",
                     "cargo:rustc-cfg=spi_master_has_octal",
                     "cargo:rustc-cfg=spi_master_has_app_interrupts",
@@ -6459,6 +6509,7 @@ pub fn emit_check_cfg_directives() {
     println!("cargo:rustc-check-cfg=cfg(soc_has_clock_node_uart_baud_rate_generator)");
     println!("cargo:rustc-check-cfg=cfg(has_dram_region)");
     println!("cargo:rustc-check-cfg=cfg(has_dram2_uninit_region)");
+    println!("cargo:rustc-check-cfg=cfg(spi_master_bit_order_is_bool)");
     println!("cargo:rustc-check-cfg=cfg(spi_master_supports_dma)");
     println!("cargo:rustc-check-cfg=cfg(spi_slave_supports_dma)");
     println!("cargo:rustc-check-cfg=cfg(timergroup_timg_has_timer1)");
@@ -6635,6 +6686,7 @@ pub fn emit_check_cfg_directives() {
     println!("cargo:rustc-check-cfg=cfg(soc_has_clock_node_parl_io_rx_clock)");
     println!("cargo:rustc-check-cfg=cfg(soc_has_clock_node_parl_io_tx_clock)");
     println!("cargo:rustc-check-cfg=cfg(spi_master_has_clk_pre_div)");
+    println!("cargo:rustc-check-cfg=cfg(spi_master_dma_can_access_flash)");
     println!("cargo:rustc-check-cfg=cfg(uart_peripheral_controls_mem_clk)");
     println!("cargo:rustc-check-cfg=cfg(uhci_combined_uart_selector_field)");
     println!("cargo:rustc-check-cfg=cfg(wifi_has_5g)");
@@ -6780,6 +6832,8 @@ pub fn emit_check_cfg_directives() {
         "cargo:rustc-check-cfg=cfg(soc_rc_fast_clk_default, \
          values(\"8500000\",\"17500000\",\"20000000\"))"
     );
+    println!("cargo:rustc-check-cfg=cfg(spi_master_version, values(\"1\",\"3\",\"2\"))");
+    println!("cargo:rustc-check-cfg=cfg(spi_master_fifo_size, values(\"64\",\"72\"))");
     println!("cargo:rustc-check-cfg=cfg(uart_ram_size, values(\"128\"))");
     println!("cargo:rustc-check-cfg=cfg(uart_version, values(\"1\",\"2\"))");
     println!("cargo:rustc-check-cfg=cfg(wifi_mac_version, values(\"1\",\"3\",\"2\"))");
