+++ esp-metadata-generated/src/_generated_esp32c61.rs
@@ -310,6 +310,21 @@ macro_rules! property {
     ("clock_tree.uart.baud_rate_generator.integral") => {
         (0, 4095)
     };
+    ("spi_master.version") => {
+        3
+    };
+    ("spi_master.version", str) => {
+        stringify!(3)
+    };
+    ("spi_master.fifo_size") => {
+        64
+    };
+    ("spi_master.fifo_size", str) => {
+        stringify!(64)
+    };
+    ("spi_master.bit_order_is_bool") => {
+        false
+    };
     ("spi_master.supports_dma") => {
         true
     };
@@ -323,6 +338,9 @@ macro_rules! property {
         true
     };
     ("spi_master.has_clk_pre_div") => {
+        true
+    };
+    ("spi_master.dma_can_access_flash") => {
         true
     };
     ("spi_slave.supports_dma") => {
