+++ esp-metadata/devices/esp32.toml
@@ -825,13 +825,15 @@ support_status = "partial"
 algo = { sha1 = 0, sha256 = 0, sha384 = 0, sha512 = 0 } # fake mode bits, ESP32 has separate register sets
 
 [device.spi_master]
+version = 1
+fifo_size = 64
+bit_order_is_bool = true
 support_status = "supported"
 supports_dma = true
 instances = [
     { name = "spi2", sys_instance = "Spi2", sclk = "HSPICLK", sio = ["HSPID", "HSPIQ", "HSPIWP", "HSPIHD"], cs = ["HSPICS0", "HSPICS1", "HSPICS2"] },
     { name = "spi3", sys_instance = "Spi3", sclk = "VSPICLK", sio = ["VSPID", "VSPIQ", "VSPIWP", "VSPIHD"], cs = ["VSPICS0", "VSPICS1", "VSPICS2"] },
 ]
-
 [device.spi_slave]
 support_status = "partial"
 supports_dma = true
