+++ esp-metadata/devices/esp32c3.toml
@@ -486,6 +486,9 @@ dma = true
 algo = { sha1 = 0, sha224 = 1, sha256 = 2 }
 
 [device.spi_master]
+version = 3
+fifo_size = 64
+bit_order_is_bool = true
 support_status = "supported"
 supports_dma = true
 has_app_interrupts = true
@@ -493,7 +496,6 @@ has_dma_segmented_transfer = true
 instances = [
     { name = "spi2", sys_instance = "Spi2", sclk = "FSPICLK", sio = ["FSPID", "FSPIQ", "FSPIWP", "FSPIHD"], cs = ["FSPICS0", "FSPICS1", "FSPICS2", "FSPICS3", "FSPICS4", "FSPICS5"] },
 ]
-
 [device.spi_slave]
 support_status = "partial"
 supports_dma = true
