+++ esp-metadata/devices/esp32c61.toml
@@ -468,6 +468,9 @@ algo = { sha1 = 0, sha224 = 1, sha256 = 2 }
 support_status = { status = "not_supported", issue = 5444 }
 
 [device.spi_master]
+dma_can_access_flash = true
+version = 3
+fifo_size = 64
 support_status = "supported"
 supports_dma = true
 has_app_interrupts = true
@@ -476,7 +479,6 @@ has_clk_pre_div = true
 instances = [
     { name = "spi2", sys_instance = "Spi2", sclk = "FSPICLK", sio = ["FSPID", "FSPIQ", "FSPIWP", "FSPIHD"], cs = ["FSPICS0", "FSPICS1", "FSPICS2", "FSPICS3", "FSPICS4", "FSPICS5"] },
 ]
-
 [device.spi_slave]
 instances = [
     { name = "spi2", sys_instance = "Spi2", sclk = "FSPICLK", mosi = "FSPID", miso = "FSPIQ", cs = "FSPICS0" },
