+++ esp-metadata/devices/esp32h2.toml
@@ -594,6 +594,8 @@ algo = { sha1 = 0, sha224 = 1, sha256 = 2 }
 support_status = { status = "not_supported", issue = 5444 }
 
 [device.spi_master]
+version = 3
+fifo_size = 64
 support_status = "supported"
 supports_dma = true
 has_app_interrupts = true
@@ -601,7 +603,6 @@ has_dma_segmented_transfer = true
 instances = [
     { name = "spi2", sys_instance = "Spi2", sclk = "FSPICLK", sio = ["FSPID", "FSPIQ", "FSPIWP", "FSPIHD"], cs = ["FSPICS0", "FSPICS1", "FSPICS2", "FSPICS3", "FSPICS4", "FSPICS5"] },
 ]
-
 [device.spi_slave]
 support_status = "partial"
 supports_dma = true
