+++ esp-metadata/devices/esp32s2.toml
@@ -610,6 +610,9 @@ dma = true
 algo = { sha1 = 0, sha224 = 1, sha256 = 2, sha384 = 3, sha512 = 4, sha512_224 = 5, sha512_256 = 6, sha512_t = 7 }
 
 [device.spi_master]
+version = 2
+fifo_size = 72
+bit_order_is_bool = true
 support_status = "supported"
 supports_dma = true
 has_octal = true
@@ -618,7 +621,6 @@ instances = [
     { name = "spi2", sys_instance = "Spi2", sclk = "FSPICLK",  sio = ["FSPID", "FSPIQ", "FSPIWP", "FSPIHD", "FSPIIO4", "FSPIIO5", "FSPIIO6", "FSPIIO7"], cs = ["FSPICS0", "FSPICS1", "FSPICS2", "FSPICS3", "FSPICS4", "FSPICS5"] },
     { name = "spi3", sys_instance = "Spi3", sclk = "SPI3_CLK", sio = ["SPI3_D", "SPI3_Q"], cs = ["SPI3_CS0", "SPI3_CS1", "SPI3_CS2"] },
 ]
-
 [device.spi_slave]
 support_status = "partial"
 supports_dma = true
