+++ esp-metadata/src/cfg.rs
@@ -663,6 +663,13 @@ driver_configs![
         driver: spi_master,
         name: "SPI master",
         properties: {
+            /// Register-layout generation derived from the chip SVD.
+            version: u32,
+            /// FIFO size in bytes.
+            fifo_size: u32,
+            /// Bit-order fields are single-bit booleans instead of multi-bit fields.
+            #[serde(default)]
+            bit_order_is_bool: bool,
             #[serde(default)]
             supports_dma: bool,
             #[serde(default)]
@@ -674,6 +681,9 @@ driver_configs![
             /// The PCR has a clock pre-divider before the SPI peripheral.
             #[serde(default)]
             has_clk_pre_div: bool,
+            /// SPI DMA can read flash memory directly.
+            #[serde(default)]
+            dma_can_access_flash: bool,
         }
     },
     SpiSlaveProperties<SpiSlaveInstanceConfig> {
